ch4/ofi: Added support to assign optimized memory regions to RMA window
The optimized memory region is a region that will be used for lower overhead,
unordered RMA operations. It will use the low-overhead RX path. Additionally,
low-overhead packet formats may be used to target an optimized MR.

In order to create an optimized memory region, MR with key values in the special
range, tentatively 0-200 will be used. Optimized MRs will not support FI_MR_RMA_EVENT.

The cxi libfabric provider will support at least 1500 optimized MRs per device.
In order to enable these memory regions, the application will need to set the
accumulate ordering to 0 as it supports unordered RMA operations.

This changeset will enable the capability to assign optimized memory regions to RMA
window operations if requested by the application. For all the other libfabric
providers, the number of optimized memory regions is set to zero in the capability set
since they do not support it.