ch4/ofi: disable the incorrect address table validation
The function get_av_table_index is returning incorrect values for the
case of  multi-nics. As a result, the validation logic during address
exchange is incorrect.

So, currently disabling the address table validation part. We need to
either correct the get_av_table_index implementation or remote these
validation logic.