maint/jenkins: Add new mpir, posix, and ch4 tuning json files
The json files are obtained using mpitune_mpich,
which performs automatic tuning of the MPI collectives.
Tuning was performed on 64 nodes of Co-val with different
algos and algo parameters, and ppn varying from 1-104.
Some of the major changes in the tuning file:

Bcast: tree algos are found to be optimal for small and
intermediate message sizes. They were not present
in the older tuning file.
Allreduce: recexch were found to be optimal
in many cases. This algo was not present in the
older tuning file.
Barrier: recexch replaced k_dissemination

mpir and ch4 json tuned on 64 nodes